// apps/inventory/src/app/remote-entry/model/areamaster.interface.ts
  export interface AREAMASTERS {
    id: number;
    code: string;
    name: string;
    isGroup: boolean;
  }

  export interface AREAMASTER {
    id: number;
    code: string;
    name: string;
    note: string;
    parentId: number;
    isGroup: boolean;
    parentName: string;
    createdBy: number;
    createdBranchId: number;
    createdOn: string; // ISO date string format
    active: boolean;
  }
  export interface AREAGROUPPOPUP {
    id: number;
    code: string;
    name: string;
  }
  

  

  